Boolean bonanza
You're not allowed to use "boolean" in the telemetry stanza, so I changed it.
For validation I needed some filter magic to check if it's a boolean or not, since
else I needed to add or remove the quotes around the value.

Also reverted the quotes around all booleans in the rest of the config template.